+++ src/redturtle/volto/tests/test_content_types.py
@@ -36,21 +36,21 @@ class TestContentTypes(unittest.TestCase):
     def test_document_folderish(self):
         portal_types = api.portal.get_tool(name="portal_types")
         self.assertEqual(
-            "collective.folderishtypes.dx.content.FolderishDocument",
+            "plone.volto.content.FolderishDocument",
             portal_types["Document"].klass,
         )
 
     def test_news_folderish(self):
         portal_types = api.portal.get_tool(name="portal_types")
         self.assertEqual(
-            "collective.folderishtypes.dx.content.FolderishNewsItem",
+            "plone.volto.content.FolderishNewsItem",
             portal_types["News Item"].klass,
         )
 
     def test_event_folderish(self):
         portal_types = api.portal.get_tool(name="portal_types")
         self.assertEqual(
-            "collective.folderishtypes.dx.content.FolderishEvent",
+            "plone.volto.content.FolderishEvent",
             portal_types["Event"].klass,
         )
 
@@ -58,13 +58,17 @@ class TestContentTypes(unittest.TestCase):
         portal_types = api.portal.get_tool(name="portal_types")
         fti = portal_types["News Item"]
         self.assertTrue(fti.filter_content_types)
-        self.assertEqual(fti.allowed_content_types, ("Image", "File", "Link"))
+        self.assertEqual(
+            fti.allowed_content_types, ("Document", "Image", "File", "Link")
+        )
 
     def test_events_can_add_only_some_types(self):
         portal_types = api.portal.get_tool(name="portal_types")
         fti = portal_types["News Item"]
         self.assertTrue(fti.filter_content_types)
-        self.assertEqual(fti.allowed_content_types, ("Image", "File", "Link"))
+        self.assertEqual(
+            fti.allowed_content_types, ("Document", "Image", "File", "Link")
+        )
 
 
 class TestContentTypesSchema(unittest.TestCase):